refactor: 매장 관리자 통계 페이지 대시보드 페이지로 통합
